requirements.txt
django
gunicorn
django-heroku
slackclient
djangorestframework